Maintain thread affinity during non-TKO failovers
Summary:
When failing over, make an educated guess as to whether thread affinity still applies to the failover tier. This is done by inspecting the failover result code. It thread affinity is likely to be broken, `SRRoute` will attempt to jump threads to restore affinity.

TKO errors should not affect thread affinity, as `traverse()` is aware of TKO hosts and should be able to predict a failover in that case:
```
FailoverRoute
  |-> traverse(): failover1
  `-> route()
       |-> primary:   TKO
       `-> failover1: SUCCESS  // <- thread affinity
```
```
FailoverRoute
  |-> traverse(): failover2
  `-> route()
       |-> primary:   TKO
       |-> failover1: TKO
       `-> failover2: SUCCESS  // <- thread affinity
```
A non-TKO error, on the other hand, will break thread affinity. For instance, during S459689:
```
FailoverRoute
  |-> traverse(): failover2
  `-> route()
       |-> primary:   TKO
       |-> failover1: BUSY  // <- original thread affinity
       `-> failover2: SUCCESS  // jump threads
```

Reviewed By: disylh, stuclar

Differential Revision: D65231378

fbshipit-source-id: be9f40aa08eb20d07374064ad99c09a63158b653