Add HashRoute salt when wrapping SRRoute for runtime var
Summary:
When using following routing tree in staging proxy with runtime var, server1 may get 100% and server2 get 0%. Because all the buckets routed to SRRoute will also be routed to one of the server when using same wch3 hash and weights.
```
   McBucketRoute
         |
   HashRoute w/ hash func WeightedCh3Rv and weight [0.5, 0.5]
      /      \
NullRoute     SRRoute
              /      \
server1 (weight 0.5) serve2 (weight 0.5)
```

The diff support adding a salt in the HashRoute when initialized by SRRoute.

Reviewed By: anidev, alfozan

Differential Revision: D78127833

fbshipit-source-id: 42da4c4df1438ecc58350b97962e249fa777bc90