Fall back to current proxy thread on affinity failures (S567996)
Summary:
When a suitable proxy thread cannot be found for a given routing key (because, say, all hosts are TKO), then we keep routing the request on the current proxy thread, instead of jumping to thread 0.

During S567996, two regions were TKO as part of a storm. Requests that would normally be spread out across all proxy threads were consequently scheduled on thread 0, which overloaded.

#buildmore

Reviewed By: disylh

Differential Revision: D83589749

fbshipit-source-id: 4ad4020780f367517493dedd93df653927966007